+++ config.example.json
@@ -97,9 +97,7 @@
     },
     "vision": {
       "enabled": true,
-      "model": "gpt-4o-mini",
-      "api_key": "env://OPENAI_API_KEY",
-      "api_base": "https://api.openai.com/v1",
+      "model_name": "gpt-4o-mini",
       "prompt": "Describe this image in detail."
     },
     "media_cleanup": {
