+++ examples/config/telegram-openrouter.json
@@ -50,8 +50,7 @@
     },
     "vision": {
       "enabled": true,
-      "model": "openrouter/z-ai/glm-5v-turbo",
-      "api_key": "env://OPENROUTER_API_KEY",
+      "model_name": "openrouter-glm-5v",
       "prompt": "Describe this image in detail."
     },
     "media_cleanup": {
